+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/ms/ckeyratios/RevenueGrowthSection.java
@@ -83,7 +83,7 @@ public class RevenueGrowthSection {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(RevenueGrowthSection.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(RevenueGrowthSection.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("tenYearAverage");
         sb.append('=');
         sb.append(((this.tenYearAverage == null)?"<null>":this.tenYearAverage));
