+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/ms/keyratios/Result.java
@@ -59,7 +59,7 @@ public class Result {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(Result.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(Result.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("periodEndDate");
         sb.append('=');
         sb.append(((this.periodEndDate == null)?"<null>":this.periodEndDate));
